requirements.txt
anyio==3.6.2
APScheduler==3.9.1.post1
async-generator==1.10
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer>=2
chromedriver-binary==77.0.3865.40.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cryptography==38.0.3
cycler==0.11.0
docopt==0.6.2
exceptiongroup==1.0.3
Flask==2.2.2
fonttools==4.38.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.23.4
opencv-python==4.6.0.66
outcome==1.2.0
packaging==21.3
Pillow==9.3.0
pipreqs==0.4.11
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-telegram-bot==20.0a4
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rfc3986==1.5.0
scipy==1.9.3
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tinydb==4.7.0
tornado==6.2
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
Werkzeug==2.2.2
webdriver-manager==3.8.5
wsproto==1.2.0
yarg==0.1.9